Time: 3 ms (50.00%) | Memory: 92.1 MB (46.55%) - LeetSync

+++ 3967-earliest-finish-time-for-land-and-water-rides-ii/earliest-finish-time-for-land-and-water-rides-ii.java
@@ -1,11 +1,11 @@
 class Solution {
-    public int solve(int FrStart[],int FsDur[],int SrStart[],int SrDur[]){
-        int min=Integer.MAX_VALUE;
+    public long solve(int FrStart[],int FsDur[],int SrStart[],int SrDur[]){
+        long min=Long.MAX_VALUE;
         for(int i=0;i<FrStart.length;i++){
             min=Math.min(min,FrStart[i]+FsDur[i]);
         }
 
-        int minTotal=Integer.MAX_VALUE;
+        long minTotal=Long.MAX_VALUE;
         for(int i=0;i<SrStart.length;i++){
             if(SrStart[i]<=min){
                 minTotal=Math.min(minTotal,min+SrDur[i]);
@@ -19,7 +19,7 @@ class Solution {
     }
     public int earliestFinishTime(int[] landStartTime, int[] landDuration, int[] waterStartTime, int[] waterDuration) {
 
-        return Math.min(solve(landStartTime,landDuration,waterStartTime,waterDuration),solve(waterStartTime,waterDuration,landStartTime,landDuration));
+        return (int)Math.min(solve(landStartTime,landDuration,waterStartTime,waterDuration),solve(waterStartTime,waterDuration,landStartTime,landDuration));
         
     }
 }
